// apps/mybookkeeper/frontend/src/shared/types/summary/month-expense-summary.ts
export interface MonthExpenseSummary {
  month: string;
  maintenance?: number;
  utilities?: number;
  management_fee?: number;
  insurance?: number;
  mortgage_interest?: number;
  mortgage_principal?: number;
  taxes?: number;
  cleaning_expense?: number;
  channel_fee?: number;
  advertising?: number;
  legal_professional?: number;
  travel?: number;
  other_expense?: number;
  contract_work?: number;
  furnishings?: number;
  supplies?: number;
  home_office?: number;
  meals?: number;
  vehicle_expenses?: number;
  health_insurance?: number;
  education_training?: number;
}